+++ tests/testthat/test-sc.R
@@ -10,3 +10,10 @@ test_that("sc accepts default names", {
   expect_equal(length(x), 2)
   expect_equal(x, sc("blue3", "pink3"))
 })
+
+
+test_that("sc_as_list() works", {
+  x <- sc_as_list("blue", "pink")
+  expect_true(is.list(x))
+  expect_equal(names(x), c("blue", "pink"))
+})