+++ NEWS.md
@@ -1,7 +1,12 @@
-# Dev
+# Development
+- Return list or named vector with new parameters
+  - `sc(..., with_names = TRUE)`
+  - `sc_across(..., return = "list")`
+  - `sc_within(..., return = "list")`
+  
 - Replace Travis CI with GitHub Actions #46
 
 # 0.1.1 Hot fix for CRAN
-- Uses `simplecolors::color_table` as namespaced default in functions #36
+- Needs namespace `simplecolors::color_table` for use in functions #36
 
 # 0.1.0 Initial release